Agile Software Development Payroll

+++ agile-software-development/payroll/src/main/java/com/genius/payroll/AddSalariedEmployee.java
@@ -4,7 +4,7 @@ public class AddSalariedEmployee extends AddEmployeeTransaction {
 
 	private double salary;
 
-	public AddSalariedEmployee(int itsEmpId, String itsName, String itsAddress, double salary) {
+	public AddSalariedEmployee(long itsEmpId, String itsName, String itsAddress, double salary) {
 		super(itsEmpId, itsName, itsAddress);
 		this.salary = salary;
 	}
